+++ google2/CHANGELOG.md
@@ -10,11 +10,12 @@ Changed
 - Add arguments to `insertBucket`
 - Fix `scala.MatchError` from `handleErrorWith`
 - Add `delete` function to `GoogleTopicAdmin` trait and implementation
+- Use `recoverWith` instead of `onError` which doesn't actually recover the error
 
 Add
 - Add `GoogleDataproc` and `GoogleDataprocInterpreter`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.6-c91d96b"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.6-TRAVIS-REPLACE-ME"`
 
 
 ## 0.5